+++ components/CounterMain.tsx
@@ -1,7 +1,8 @@
 import styles from './CounterMain.module.css'
+import React from "react";
 
 
-const CounterMain = ({children}: {children: JSX.Element[]}) => {
+const CounterMain = ({children}: {children: React.JSX.Element[]}) => {
   return (
     <div className={styles.main}>
       {children}