+++ components/LinkKeyCounter.tsx
@@ -1,0 +1,8 @@
+import Link from "next/link";
+import React from "react";
+
+export function LinkKeyCounter() {
+  return (
+    <Link href={{pathname: '/key-refined-counter'}}>TF2 Key and Refined Counter</Link>
+  )
+}