+++ components/LinkSplackDiscord.tsx
@@ -1,0 +1,8 @@
+import Link from "next/link";
+import React from "react";
+
+export function LinkSplackDiscord() {
+  return (
+    <a href="https://discord.gg/BSSrHRd" title="House Of Spells: A TF2 Spell Trading Community">SPLACK Discord</a>
+  )
+}