+++ CHANGES.md
@@ -1,5 +1,13 @@
 # There are the changes to the project.
 
+## v1.4.0
+
+Date: 11/05/2023
+
+    * make spell count links more prominent on home page
+    * fix missing The Superfan and The Athletic Supporter
+    * add link to "House Of Spells: A TF2 Spell Trading Community": [SPLACK Discord](https://discord.gg/BSSrHRd)
+
 ## v1.3.0
 
 Date: 08/06/2023